feat(Mybatis):test

+++ src/main/java/qin/com/entity/Users.java
@@ -1,15 +1,17 @@
 package qin.com.entity;
 
 public class Users {
-    private Integer id;
+    private Long id;
+
     private String name;
+
     private String pass;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
